buttonReward

+++ train_src/madrona_escape_room_learn/rnn.py
@@ -25,10 +25,9 @@ class LSTM(nn.Module):
         self.hidden_shape = (2, self.num_layers, hidden_channels)
 
     def forward(self, in_features, cur_hidden):
-        in_features = in_features.view(1, *in_features.shape)
-
-        out, (new_h, new_c) = self.lstm(in_features,
-                                        (cur_hidden[0], cur_hidden[1]))
+        in_features = in_features.view(1, *in_features.shape).half()
+        
+        out, (new_h, new_c) = self.lstm(in_features, (cur_hidden[0], cur_hidden[1]))
 
         new_hidden = torch.stack([new_h, new_c], dim=0)
